requirements.txt
numpy>=1.24
Pillow>=9.5
torch>=2.0
matplotlib>=3.7
